backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.10
numpy>=1.22.0
pillow>=9.0.0
face_recognition>=1.3.0
pymongo>=4.0.0
dlib==19.24.8
python-dotenv>=0.19.0
opencv-python>=4.5.0
